pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bytemymood"
version = "0.1.0"
description = "A multi-agent recipe planning system that helps you discover, plan, and prepare meals based on your preferences and available ingredients."
authors = [
    {name = "Ling-Yuan Chen", email = "alanchenly@gmail.com"}
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0,<2.0.0",
    "pydantic>=2.10.6,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "google-genai>=1.16.1,<2.0.0",
    "google-adk>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5,<9.0.0",
    "google-adk[eval]>=1.0.0,<2.0.0",
    "pytest-asyncio>=0.26.0,<0.27.0",
]
deployment = [
    "absl-py>=2.2.1,<3.0.0",
    "cloudpickle>=3.1.1,<4.0.0",
    "flake8-pyproject>=1.2.3,<2.0.0",
]

[tool.setuptools]
packages = ["bytemymood"] 